requirements.txt
python-telegram-bot[httpx]==20.0